Update price field to use BigDecimal
Changed price data type from Double to BigDecimal for better precision.

+++ src/main/java/com/abml/jpa/hibernate/model/Product.java
@@ -91,11 +91,11 @@ public class Product {
         this.description = description;
     }
 
-    public Double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(Double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
